Support for Win32 platform added (Visual Studio 2003 for now)
git-svn-id: https://klub.com.pl/var/svn/omnet@210 056025b2-ba0e-0410-831a-f040b6dfa49e

+++ util/logger.h
@@ -21,7 +21,7 @@
 #include <string>
 
 #define Log(X) SLog(this,X)
-#define SLog(OBJ,X) logger::setSimTime(OBJ->simTime()); logger::setLogName(OBJ->fullName()); logger::setEv(&ev); logger :: log##X ()
+#define SLog(OBJ,X) logger::setSimTime((float)OBJ->simTime()); logger::setLogName(OBJ->fullName()); logger::setEv(&ev); logger :: log##X ()
 #define LogEnd logger :: endl
 
 #define LOGMODE_DEFAULT LOGMODE_SIMTIME
